requirements.txt
flask==3.0.3
gunicorn==21.2.0
numpy==1.26.4
pandas==1.5.3
scikit-learn==1.5.2
tensorflow-cpu==2.13.0
